Move version-agnostic welcome to routes index
