+++ application/api/routes/__init__.py
@@ -1,6 +1,13 @@
 # coding: utf-8
 
-from .misc import ApiWelcome
+from .handler import BaseHandler
+class ApiWelcome(BaseHandler):
+    def get(self):
+        self.jsonify(
+            success=True,
+            message='Edaemon API welcomes you',
+            current_version=1
+        )
 
 all_routes = [
     (r'/api/', ApiWelcome)
