--- application/api/routes/misc.py
@@ -1,11 +1,0 @@
-# coding: utf-8
-
-from .handler import BaseHandler
-
-class ApiWelcome(BaseHandler):
-    def get(self):
-        self.jsonify(
-            success=True,
-            message='Edaemon API welcomes you',
-            current_version=1
-        )
